frontend/requirements.txt
flask
google-api-core
google-cloud-tasks
google-cloud-storage
zulu
requests
